requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
arrow==0.15.5
astroid>=2.2.0
attrs==18.2.0
backcall==0.1.0
beautifulsoup4==4.7.1
black==20.8b1
bleach==3.1.0
cachetools==4.1.1
certifi==2018.11.29
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cssselect==1.1.0
cssutils==1.0.2
dataclasses==0.7
decorator==4.4.2
defusedxml==0.5.0
Deprecated==1.2.5
Django==2.1.5
django-cors-headers==3.2.0
django-extensions==2.1.6
django-redis==4.10.0
django-rq==1.3.0
django-webpack-loader==0.6.0
entrypoints==0.3
feedfinder2==0.0.4
feedparser==5.2.1
feedsearch==1.0.12
filelock==3.0.12
Flask==1.1.2
furl==2.1.0
google==2.0.1
gunicorn==19.9.0
guppy3==3.0.9
html2text==2018.1.9
idna==2.8
importlib-metadata==4.6.3
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.9
itsdangerous==1.1.0
jedi==0.13.2
jieba3k==0.35.1
joblib==0.17.0
jsonschema==3.0.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
lxml==4.5.2
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbconvert==5.4.1
nbformat==4.4.0
newspaper3k==0.2.8
nltk==3.5
notebook==5.7.4
oauthlib==3.0.1
orderedmultidict==1.0.1
pandocfilters==1.4.2
parso==0.3.3
pathspec==0.8.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.0.1
pocket==0.3.6
polyglot==16.7.4
premailer==3.7.0
prometheus-client==0.6.0
prompt-toolkit==2.0.8
psycopg2==2.8.1
ptyprocess==0.6.0
pycparser==2.20
PyGithub==1.43.5
Pygments==2.3.1
PyICU==2.3.1
PyJWT==2.1.0
pylint==2.3.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyrsistent==0.14.11
PySocks==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.3.1
pyzmq==18.0.0
qtconsole==4.4.3
redis==3.2.0
regex==2020.10.23
requests==2.21.0
requests-file==1.5.1
requests-oauthlib==1.2.0
rq==0.13.0
Send2Trash==1.5.0
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==1.8
terminado==0.8.1
testpath==0.4.2
tinysegmenter==0.3
tldextract==3.0.2
toml==0.10.1
tornado==5.1.1
tqdm==4.51.0
traitlets==4.3.2
tweepy==3.10.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.4.2
wrapt==1.11.1
zipp==3.5.0
